fix: moved on click handler which triggers filetype handling
The on click handler which was responsible for changing the filetype with
a telescope window was moved from the default file_info function to the
statusline file info table.
The reason for that is that it triggered the event both on the status line
and the winbar which overrides the winbar tab behaviour when clicked with
the mouse.